+++ test/test_memory_layout.cc
@@ -32,7 +32,7 @@ BOOST_AUTO_TEST_CASE( test_layout_simple )
         MemoryLayout ops = Typelib::layout_of(type);
 
         BOOST_REQUIRE_EQUAL(2U, ops.size());
-        BOOST_REQUIRE_EQUAL(MemLayout::FLAG_MEMCPY, ops[0]);
+        BOOST_REQUIRE_EQUAL((size_t)MemLayout::FLAG_MEMCPY, ops[0]);
         A data;
         BOOST_REQUIRE_EQUAL(offsetof(A, d) + sizeof(data.d), ops[1]);
     }
@@ -43,7 +43,7 @@ BOOST_AUTO_TEST_CASE( test_layout_simple )
         MemoryLayout ops = Typelib::layout_of(type);
 
         BOOST_REQUIRE_EQUAL(2U, ops.size());
-        BOOST_REQUIRE_EQUAL(MemLayout::FLAG_MEMCPY, ops[0]);
+        BOOST_REQUIRE_EQUAL((size_t)MemLayout::FLAG_MEMCPY, ops[0]);
         BOOST_REQUIRE_EQUAL(100U, ops[1]);
     }
 
@@ -53,7 +53,7 @@ BOOST_AUTO_TEST_CASE( test_layout_simple )
         MemoryLayout ops = Typelib::layout_of(type);
 
         BOOST_REQUIRE_EQUAL(2U, ops.size());
-        BOOST_REQUIRE_EQUAL(MemLayout::FLAG_MEMCPY, ops[0]);
+        BOOST_REQUIRE_EQUAL((size_t)MemLayout::FLAG_MEMCPY, ops[0]);
         B data;
         BOOST_REQUIRE_EQUAL(offsetof(B, z) + sizeof(data.z), ops[1]);
     }
@@ -64,7 +64,7 @@ BOOST_AUTO_TEST_CASE( test_layout_simple )
         MemoryLayout ops = Typelib::layout_of(type);
 
         BOOST_REQUIRE_EQUAL(2U, ops.size());
-        BOOST_REQUIRE_EQUAL(MemLayout::FLAG_MEMCPY, ops[0]);
+        BOOST_REQUIRE_EQUAL((size_t)MemLayout::FLAG_MEMCPY, ops[0]);
         BOOST_REQUIRE_EQUAL(type.getSize(), ops[1]);
     }
 
@@ -75,8 +75,8 @@ BOOST_AUTO_TEST_CASE( test_layout_simple )
         BOOST_CHECK_THROW(Typelib::layout_of(type), Typelib::NoLayout);
 
         MemoryLayout ops = Typelib::layout_of(type, true);
-        BOOST_REQUIRE_EQUAL(2, ops.size());
-        BOOST_REQUIRE_EQUAL(MemLayout::FLAG_MEMCPY, ops[0]);
+        BOOST_REQUIRE_EQUAL(2U, ops.size());
+        BOOST_REQUIRE_EQUAL((size_t)MemLayout::FLAG_MEMCPY, ops[0]);
         C data;
         BOOST_REQUIRE_EQUAL(offsetof(C, z) + sizeof(data.z), ops[1]);
     }
